refactor: KeyValueInput improvements
