+++ src/components/atoms/KeyValueInput/KeyValueEntryRenderer.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 
-import {Input, Select} from 'antd';
+import {Select} from 'antd';
 
 import {MinusOutlined} from '@ant-design/icons';
 
 import Colors from '@styles/Colors';
 
-import {ANY_VALUE} from './constants';
+import ValueInput from './ValueInput';
 import {KeyValueEntry} from './types';
 
 import * as S from './styled';
@@ -20,39 +20,6 @@ type KeyValueEntryRendererProps = {
   disabled?: boolean;
   availableKeys: string[];
   availableValues?: string[];
-};
-
-type ValueInputProps = {
-  value?: string;
-  valueType: string;
-  availableValues?: string[];
-  onChange: (newValue: string) => void;
-  disabled?: boolean;
-};
-
-const ValueInput: React.FC<ValueInputProps> = props => {
-  const {value, valueType, availableValues, disabled, onChange} = props;
-
-  if (valueType === 'string') {
-    if (availableValues?.length) {
-      return (
-        <Select value={value} onChange={onChange} showSearch disabled={disabled}>
-          <Select.Option key={ANY_VALUE} value={ANY_VALUE}>
-            {ANY_VALUE}
-          </Select.Option>
-          {availableValues?.map((valueOption: string) => (
-            <Select.Option key={valueOption} value={valueOption}>
-              {valueOption}
-            </Select.Option>
-          ))}
-        </Select>
-      );
-    }
-    return <Input value={value} onChange={e => onChange(e.target.value)} disabled={disabled} />;
-  }
-
-  // TODO: decide if we want to implement more value types
-  return null;
 };
 
 const KeyValueEntryRenderer: React.FC<KeyValueEntryRendererProps> = props => {
